requirements.txt
Flask>=2.0.0
requests>=2.25.0
protobuf>=3.19.0
pycryptodome>=3.15.0
urllib3>=1.26.0

